split web models and fields
