+++ leonardo/module/web/fields/common.py
@@ -1,0 +1,8 @@
+
+from django_select2.fields import *
+
+from django_select2.widgets import *
+
+
+class Field(AutoModelSelect2Field):
+    empty_values = [u'']
