+++ leonardo/module/web/models/__init__.py
@@ -1,0 +1,3 @@
+
+from leonardo.module.web.models.page import *
+from leonardo.module.web.models.widget import *
